split heatmap into components
